Fix all type checking issues.
