+++ lib/package/audit/const/yaml.rb
@@ -1,0 +1,9 @@
+module Package
+  module Audit
+    module Const
+      module YAML
+        TECHNOLOGY = 'technology'
+      end
+    end
+  end
+end
