+++ lib/package/audit/enum/technology.rb
@@ -4,6 +4,10 @@ module Package
       module Technology
         NODE = 'node'
         RUBY = 'ruby'
+
+        def self.all
+          constants.map { |key| Enum::Technology.const_get(key) }.sort
+        end
       end
     end
   end
