+++ lib/package/audit/formatter/vulnerability.rb
@@ -1,6 +1,6 @@
-require_relative 'base'
 require_relative '../enum/vulnerability_type'
 require_relative '../util/bash_color'
+require_relative 'base'
 
 module Package
   module Audit
