--- sig/package/audit/command_parser.rbs
@@ -1,13 +1,0 @@
-module Package
-  module Audit
-    class CommandParser
-      @config: Hash[String, untyped]
-      @dir: String
-      @options: Hash[Symbol, untyped]
-      @report: Symbol
-      @technologies: Array[String]
-
-      def initialize: (Symbol, String, Hash[Symbol, untyped]) -> void
-    end
-  end
-end
