+++ sig/package/audit/enum/option.rbs
@@ -1,0 +1,13 @@
+module Package
+  module Audit
+    module Enum
+      module Option
+        CONFIG: String
+        CSV: String
+        CSV_EXCLUDE_HEADERS: String
+        INCLUDE_IGNORED: String
+        TECHNOLOGY: String
+      end
+    end
+  end
+end
