+++ sig/package/audit/enum/technology.rbs
@@ -1,0 +1,12 @@
+module Package
+  module Audit
+    module Enum
+      module Technology
+        NODE: String
+        RUBY: String
+
+        def self.all: -> Array[String]
+      end
+    end
+  end
+end
