+++ sig/package/audit/technology/detector.rbs
@@ -1,0 +1,19 @@
+module Package
+  module Audit
+    module Technology
+      class Detector
+        @dir: String
+
+        def initialize: (String) -> void
+
+        def detect: -> Array[String]
+
+        private
+
+        def node?: -> bool
+
+        def ruby?: -> bool
+      end
+    end
+  end
+end
