+++ sig/package/audit/technology/validator.rbs
@@ -1,0 +1,19 @@
+module Package
+  module Audit
+    module Technology
+      class Validator
+        @dir: String
+
+        def initialize: (String) -> void
+
+        def validate!: (String) -> void
+
+        private
+
+        def validate_node!: -> void
+
+        def validate_ruby!: -> void
+      end
+    end
+  end
+end
