+++ test/package/audit/test_risk.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 
-require_relative '../../../lib/package/audit/risk'
+require_relative '../../../lib/package/audit/models/risk'
 require_relative '../../../lib/package/audit/enum/risk_type'
 
 module Package
